(clustering) move reduction logic to weekly (#2190)
