+++ jobs/ml_jobs/clusterisation/configs/default-config.json
@@ -88,5 +88,45 @@
                 }
             ]
         }
-    ]
+    ],
+    "reduction_configs": {
+        "default": [
+            {
+                "target_dimensions": 64,
+                "method": "PCA",
+                "embedding_columns": [
+                    "semantic_content_hybrid_embedding"
+                ]
+            },
+            {
+                "target_dimensions": 32,
+                "method": "PCA",
+                "embedding_columns": [
+                    "image_embedding",
+                    "semantic_content_embedding",
+                    "semantic_content_hybrid_embedding"
+                ]
+            },
+            {
+                "target_dimensions": 16,
+                "method": "PUMAP",
+                "embedding_columns": [
+                    "image_embedding",
+                    "semantic_content_embedding",
+                    "semantic_content_hybrid_embedding"
+                ],
+                "pca_pre_reduction_dimension" : 32
+            },
+            {
+                "target_dimensions": 5,
+                "method": "UMAP",
+                "embedding_columns": [
+                    "image_embedding",
+                    "semantic_content_embedding",
+                    "semantic_content_hybrid_embedding"
+                ],
+                "pca_pre_reduction_dimension" : 16
+            }
+        ]
+    }
 }
