+++ jobs/ml_jobs/clusterisation/requirements.txt
@@ -7,15 +7,17 @@ gcsfs==2023.5.0
 pandas-gbq>=0.17.1
 numpy>=1.23.3 
 hdbscan>=0.8.28
-umap>=0.1.1
 umap_learn>=0.5.3
 loguru==0.7.0
 typer==0.7.0
 pyarrow==13.0.0
-polars==0.19.12
+polars==0.18.6
 timeout-decorator
 tensorflow
 polyfuzz
 bertopic[spacy]
 flair
 openai
+umap-learn[parametric_umap]
+numba
+pynndescent
