+++ jobs/ml_jobs/clusterisation/tools/utils.py
@@ -51,11 +51,19 @@ def load_config_file(config_file_name):
         return json.load(config_file)
 
 
-def convert_str_emb_to_float(emb_list, emb_size=5):
+def convert_str_emb_to_float(emb_list):
     float_emb = []
     for str_emb in emb_list:
         emb = json.loads(str_emb)
         float_emb.append(np.array(emb))
+    return float_emb
+
+
+def convert_arr_emb_to_str(emb_list):
+    float_emb = []
+    for str_emb in emb_list:
+        emb = json.dumps(str_emb.tolist())
+        float_emb.append(emb)
     return float_emb
