+++ jobs/ml_jobs/embeddings/requirements.txt
@@ -13,6 +13,3 @@ requests==2.31.0
 gcsfs>= 2023.5.0
 protobuf>=3.9.2
 polars==0.18.6
-umap-learn[parametric_umap]
-numba
-pynndescent
