+++ orchestration/dags/jobs/ml/clusterisation_items.py
@@ -15,9 +15,21 @@ from dependencies.ml.clusterisation.import_data import (
     IMPORT_ITEM_EMBEDDINGS,
 )
 
+from airflow.providers.google.cloud.operators.bigquery import BigQueryInsertJobOperator
+from airflow.providers.google.cloud.operators.bigquery import (
+    BigQueryExecuteQueryOperator,
+    BigQueryInsertJobOperator,
+)
+from jobs.ml.constants import IMPORT_TRAINING_SQL_PATH
+from common.config import (
+    GCP_PROJECT_ID,
+    DAG_FOLDER,
+    ENV_SHORT_NAME,
+    MLFLOW_BUCKET_NAME,
+    BIGQUERY_TMP_DATASET,
+)
 
 from common.alerts import task_fail_slack_alert
-from common.config import DAG_FOLDER, ENV_SHORT_NAME
 from common.utils import get_airflow_schedule
 
 DEFAULT_REGION = "europe-west1"
@@ -30,6 +42,9 @@ default_args = {
     "on_failure_callback": task_fail_slack_alert,
     "retries": 0,
     "retry_delay": timedelta(minutes=2),
+}
+dag_config = {
+    "STORAGE_PATH": f"gs://{MLFLOW_BUCKET_NAME}/clusterisation_items_{ENV_SHORT_NAME}/clusterisation_items_{DATE}/{DATE}_item_embbedding_data",
 }
 
 
@@ -85,6 +100,46 @@ with DAG(
         instance_name=GCE_INSTANCE,
         base_dir=BASE_DIR,
         command="""pip install -r requirements.txt --user && python3 -m spacy download fr_core_news_sm""",
+    )
+
+    export_task = BigQueryExecuteQueryOperator(
+        task_id=f"import_item_embbedding_data",
+        sql=(IMPORT_TRAINING_SQL_PATH / f"item_embeddings_reduction.sql").as_posix(),
+        write_disposition="WRITE_TRUNCATE",
+        use_legacy_sql=False,
+        destination_dataset_table=f"{BIGQUERY_TMP_DATASET}.{DATE}_item_embeddings_reduction",
+        dag=dag,
+    )
+
+    export_bq = BigQueryInsertJobOperator(
+        task_id=f"store_item_embbedding_data",
+        configuration={
+            "extract": {
+                "sourceTable": {
+                    "projectId": GCP_PROJECT_ID,
+                    "datasetId": BIGQUERY_TMP_DATASET,
+                    "tableId": f"{DATE}_item_embeddings_reduction",
+                },
+                "compression": None,
+                "destinationUris": f"{dag_config['STORAGE_PATH']}/data-*.parquet",
+                "destinationFormat": "PARQUET",
+            }
+        },
+        dag=dag,
+    )
+
+    reduce_dimension = SSHGCEOperator(
+        task_id="reduce_dimension",
+        instance_name=GCE_INSTANCE,
+        base_dir=BASE_DIR,
+        environment=dag_config,
+        command="PYTHONPATH=. python reduction/dimension_reduction.py "
+        f"--gcp-project {GCP_PROJECT_ID} "
+        f"--env-short-name {ENV_SHORT_NAME} "
+        "--config-file-name {{ params.config_file_name }} "
+        f"--source-gs-path {dag_config['STORAGE_PATH']} "
+        f"--output-table-name item_embeddings "
+        f"--reduction-config default ",
     )
 
     preprocess_clustering = SSHGCEOperator(
@@ -143,6 +198,9 @@ with DAG(
         >> bq_import_item_embeddings_task
         >> gce_instance_start
         >> fetch_code
+        >> export_task
+        >> export_bq
+        >> reduce_dimension
         >> install_dependencies
         >> preprocess_clustering
         >> generate_clustering
